# packages/api-contract/specs/swagger-app.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/notification": {
      "post": {
        "operationId": "NotificationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create notification",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      },
      "get": {
        "operationId": "NotificationController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SYSTEM_ANNOUNCEMENT",
                "GENERAL"
              ]
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List notifications (pagination + filters)",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/device-token": {
      "post": {
        "operationId": "NotificationController_registerDeviceToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device token registered successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register current device FCM token",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/device-token/remove": {
      "post": {
        "operationId": "NotificationController_removeDeviceToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveDeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device token removed successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove current device FCM token",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/me": {
      "get": {
        "operationId": "NotificationController_findMine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SYSTEM_ANNOUNCEMENT",
                "GENERAL"
              ]
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Load notifications of current user",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/unread": {
      "get": {
        "operationId": "NotificationController_unread",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unread notifications counted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Count unread notifications of current user",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/read": {
      "post": {
        "operationId": "NotificationController_readNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification marked as read successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark one notification as read",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/read-all": {
      "post": {
        "operationId": "NotificationController_readAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All notifications marked as read successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read for current user",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/{id}": {
      "get": {
        "operationId": "NotificationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get notification by id",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      },
      "patch": {
        "operationId": "NotificationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update notification",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      },
      "delete": {
        "operationId": "NotificationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete notification (soft)",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/notification/{id}/restore": {
      "patch": {
        "operationId": "NotificationController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore soft-deleted notification",
        "tags": [
          "Notification"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
...
    "/api/v1/reward/daily": {
      "post": {
        "operationId": "RewardController_claimDaily",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Reward claimed, remaining actions returned"
          },
          "400": {
            "description": "Daily reward already claimed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim daily reward (free bonus action)",
        "tags": [
          "Reward"
        ],
        "x-scope": [
          "app"
        ]
      }
    },
    "/api/v1/reward/ads": {
      "post": {
        "operationId": "RewardController_rewardAds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ad reward granted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Earn bonus action from watching an ad",
        "tags": [
          "Reward"
        ],
        "x-scope": [
          "app"
        ]
      }
    },
    "/api/v1/reward-chat/ads": {
      "post": {
        "operationId": "RewardChatController_rewardAds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Chat ad reward granted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Earn bonus AI chat turns from watching an ad",
        "tags": [
          "Reward Chat"
        ],
        "x-scope": [
          "app"
        ]
      }
    },
...
  },
  "info": {
    "title": "PP02 API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token (Bearer <your_token_here>)"
      }
    },
    "schemas": {
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM_ANNOUNCEMENT",
              "GENERAL"
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "referenceId": {
            "type": "string"
          },
          "referenceType": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "message"
        ]
      },
      "RegisterDeviceTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Firebase Cloud Messaging device token",
            "example": "eXampleFcmTokenFromMobileApp"
          }
        },
        "required": [
          "token"
        ]
      },
      "RemoveDeviceTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Firebase Cloud Messaging device token to remove",
            "example": "eXampleFcmTokenFromMobileApp"
          }
        },
        "required": [
          "token"
        ]
      },
      "ReadNotificationDto": {
        "type": "object",
        "properties": {
          "idNotification": {
            "type": "string",
            "description": "Notification id",
            "format": "uuid"
          }
        },
        "required": [
          "idNotification"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM_ANNOUNCEMENT",
              "GENERAL"
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "referenceId": {
            "type": "string"
          },
          "referenceType": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}